+++ test/Kista.Manager.Events.XUnit/GlobalUsings.cs
@@ -1,0 +1,7 @@
+global using Bogus;
+global using Xunit;
+global using Microsoft.Extensions.DependencyInjection;
+global using Kista;
+global using Kista.Events;
+global using Deveel;
+global using NSubstitute;